Refactor Ember.platform.
Split various polyfils into separate files.  This will ultimately be
pulled into 1.8 beta series.
